refactor: extract implementations from mod.rs files (#79)
mod.rs files should contain only module declarations and re-exports.

service/orchestrator/mod.rs (720 → 51 lines):
- Extracted ServiceOrchestrator struct + impl to core.rs
- Extracted PrefetchHandle, MountResult, StartupProgress,
  StartupResult to types.rs

aircraft_position/web_api/mod.rs (27 → 14 lines):
- Moved SharedSimState type alias and factory to sim_state.rs
